Internal bundles: require flow_selection
It's not well specified which internal flows should be included if you
have a Bundle starting and ending in the same place; probably including
all of them is not what you want. Internal flow bundles are allowed now
only if they have a flow_selection picking out the internal flows you
really want.