Adding meta.yaml file required for building Conda
The commented out section shows the requirements and test from the
tutorial on how to build a conda package from scratch. I wasn't sure what
needs to go there. Rest of the file should be good to go. I should be able
to build the conda packages presuming the given information is correct.