Add options `add_elsewhere_waypoints` to weave()
When True (default), keep previous behaviour of adding Waypoints to
Elsewhere bundles (resulting in flows indicated by left- or
right-pointing arrows). When False, these flows are instead shown by
short "stubs" on the nodes themselves, reducing the number of flows on
the diagram.

Depends on not-yet-released additions to d3-sankey-diagram.